refactor(core,sdks): unify Retry into a single shared Context.Service
Per-SDK Retry tags were never actually consumed: each
`packages/<sdk>/src/retry.ts` declared its own `Retry extends
Context.Service<...>("FooRetry")`, but the shared `makeAPI` in
`@distilled.cloud/core/client` had no way to read from N different tags
and so just ran a hard-coded throttling-only schedule. The previous
commit on this branch threaded a per-SDK tag through `makeAPI` config —
that worked but it left every other SDK on the legacy throttling-only
path until they migrated.

Drop the per-SDK indirection entirely:

- `packages/core/src/retry.ts` now exports a single shared `Retry`
  Context.Service plus `policy`, `none`, `throttling`, `transient`
  helpers bound to it. `makeRetryService` is kept as a deprecated
  no-op shim for any external callers.
- `packages/core/src/client.ts`'s `makeAPI` always reads the shared
  `Retry` with `Effect.serviceOption`, falling back to
  `Retry.makeDefault` (transient/throttling/server with capped
  exponential backoff + jitter, 5 attempts). The `config.retry` field
  introduced in the previous commit and the legacy hard-coded
  throttling-only branch are removed. This mirrors AWS's pattern in
  `packages/aws/src/client/api.ts`.
- All 17 per-SDK `retry.ts` files (cloudflare, turso, planetscale,
  neon, stripe, kubernetes, axiom, posthog, gcp, coinbase, supabase,
  expo-eas, fly-io, mongodb-atlas, typesense, workos, prisma-postgres,
  azure) become thin re-exports of the shared core module so existing
  call sites (`Cloudflare.Retry.transient(eff)`,
  `Layer.succeed(Turso.Retry.Retry, customPolicy)`, etc.) keep working
  but now actually take effect.

AWS continues to use its own client and its own `Retry` tag —
unchanged in this PR.
